Feat: Organisms/Form dynamic options for select, multiselect
Fix: Adjust amount display in read-only (removing left and right extra padding)
