+++ src/assets/_amount.scss
@@ -21,6 +21,7 @@
     text-align: left;
     border: none;
     height: unset;
+    padding: 6px 0px;
   }
 
   &.amount-field-error {
